added sync interface for scanning sent transactions
