+++ zingo-sync/src/sync.rs
@@ -3,7 +3,7 @@
 use std::collections::{BTreeMap, HashMap};
 use std::sync::atomic::{self, AtomicBool};
 use std::sync::Arc;
-use std::time::Duration;
+use std::time::{Duration, SystemTime};
 
 use tokio::sync::{mpsc, Mutex};
 
@@ -271,6 +271,8 @@ pub fn scan_pending_transaction<W>(
     wallet: &mut W,
     transaction: Transaction,
     status: ConfirmationStatus,
+    datetime: u32,
+    price: Option<f64>,
 ) where
     W: SyncWallet + SyncBlocks + SyncTransactions + SyncNullifiers + SyncOutPoints,
 {
@@ -295,6 +297,8 @@ pub fn scan_pending_transaction<W>(
         &mut pending_transaction_nullifiers,
         &mut pending_transaction_outpoints,
         &transparent_addresses,
+        datetime,
+        price,
     )
     .unwrap();
 
@@ -469,6 +473,11 @@ async fn process_mempool_transaction<W>(
         wallet,
         transaction,
         ConfirmationStatus::Mempool(block_height),
+        SystemTime::now()
+            .duration_since(SystemTime::UNIX_EPOCH)
+            .unwrap()
+            .as_secs() as u32,
+        None,
     );
 
     // TODO: consider logic for pending spent being set back to None when txs are evicted / never make it on chain
